GITHUB: Added windows compatibility
Added build.bat and develop.bat files. Fixed bugs in c++ files that
ruined compatibility between windows and linux, so long types are now
int64_t. The tr1 namespace was removed for modern c++11. Followed all
other fixes here except for ones that modify core libraries:
https://github.com/facebookresearch/SparseConvNet/issues/128

+++ sparseconvnet/SCN/CUDA/BatchNormalization.cu
@@ -46,7 +46,7 @@ __global__ void BatchNormalization_f_train(
       runningVar[plane] = momentum * runningVar[plane] +
                           (1 - momentum) * _saveInvStd / (nActive - 1);
     }
-    _saveInvStd = pow(_saveInvStd / nActive + eps, -0.5);
+    _saveInvStd = pow(double(_saveInvStd / nActive + eps), -0.5);
     if (threadIdx.y == 0)
       saveInvStd[plane] = _saveInvStd;
     __syncthreads();
